+++ example/main.dart
@@ -2,7 +2,7 @@ import 'package:local_case/local_case.dart';
 import 'package:local_case/enum/local_case_language_enum.dart';
 
 void main() {
-  // Example of converting text to lowercase in different languages
+  // Farklı dillerde metni küçük harfe dönüştürme örnekleri
   print(LocalCase.toLowercase(
       text: 'ÉNERGIE RENOUVELABLE', language: LocalCaseLanguage.french));
   print(LocalCase.toLowercase(
@@ -10,35 +10,42 @@ void main() {
   print(LocalCase.toLowercase(
       text: 'GUTEN TAG', language: LocalCaseLanguage.german));
 
-  // Example of converting text to uppercase in Turkish
+  // Türkçe büyük harfe dönüştürme örneği
   print(LocalCase.toUppercase(
       text: 'merhaba dünya', language: LocalCaseLanguage.turkish));
 
-  // Example of converting a number to words in different languages
+  // Farklı dillerde sayıları kelimelere dönüştürme örnekleri
   print(LocalCase.convertNumberToWords(
-      number: 123, language: LocalCaseLanguage.german)); // Output: "einhundertdreiundzwanzig"
+      number: 123,
+      language: LocalCaseLanguage.german)); // Çıktı: "einhundertdreiundzwanzig"
   print(LocalCase.convertNumberToWords(
-      number: 456, language: LocalCaseLanguage.spanish)); // Output: "cuatrocientos cincuenta y seis"
+      number: 456,
+      language: LocalCaseLanguage
+          .spanish)); // Çıktı: "cuatrocientos cincuenta y seis"
   print(LocalCase.convertNumberToWords(
-      number: 789, language: LocalCaseLanguage.french)); // Output: "sept cent quatre-vingt-neuf"
+      number: 789,
+      language:
+          LocalCaseLanguage.french)); // Çıktı: "sept cent quatre-vingt-neuf"
   print(LocalCase.convertNumberToWords(
-      number: 234, language: LocalCaseLanguage.czech)); // Output: "dvě stě třicet čtyři"
+      number: 234,
+      language: LocalCaseLanguage.czech)); // Çıktı: "dvě stě třicet čtyři"
   print(LocalCase.convertNumberToWords(
-      number: 567, language: LocalCaseLanguage.turkish)); // Output: "beş yüz altmış yedi"
+      number: 567,
+      language: LocalCaseLanguage.turkish)); // Çıktı: "beş yüz altmış yedi"
 
-  // Error handling example
+  // Hata yönetimi örneği
   try {
     print(LocalCase.toLowercase(
         text: 'unsupported', language: LocalCaseLanguage.unknown));
   } catch (e) {
-    print(e); // Output: LocaleCaseError: Unsupported language
+    print(e); // Çıktı: LocaleCaseError: Unsupported language
   }
 
-  // Example of error handling with invalid number range
+  // Geçersiz sayı aralığı için hata yönetimi örneği
   try {
     print(LocalCase.convertNumberToWords(
         number: 1000000, language: LocalCaseLanguage.turkish));
   } catch (e) {
-    print(e); // Output: LocaleCaseError: Invalid range
+    print(e); // Çıktı: LocaleCaseError: Invalid range
   }
 }